requirements.txt
fastapi
uvicorn
jinja2
python-multipart 
llama-index
gpt-index
openai
pypdf
python-dotenv